harness: count per-message persistence in dm-loss, and share the gap-shape rule
dm-loss counts frames. It also tallied posts decrypted, but as a running
total - so it could never say WHICH message went missing, and on
multi-device accounts it counted self-sync copies too. That is why it
reported 201/201 and 0% on the canonical accounts while the operator
watched those same accounts' other devices receive ~10 of 200.

So persistence is now counted per message number, with the same
contiguous-tail-vs-scattered shape report dm-multidevice uses. Those
helpers move to persistence.ts and both scenarios import them, because
two copies of the gap-shape rule silently disagreeing is exactly the
drift that costs a day.

This opens the one configuration nothing has measured: per-message
persistence on aged real accounts. Frame arrival there is known clean;
whether the app keeps the messages has never been asked.
