+++ src/dev/tests/harness/dm-multidevice.scenario.test.ts
@@ -38,6 +38,7 @@ import { config } from './env';
 import { direction, subscribedInboxes } from './loss';
 import { installLockProbe, summariseLockHolds } from './lock-probe';
 import { deleteFault, makeApiClient } from './transport';
+import { missingReport, postsMatching } from './persistence';
 import { RunLog } from './log';
 
 const ROUNDS = Number(process.env.HARNESS_MD_ROUNDS ?? 20);
@@ -56,47 +57,6 @@ const SETTLE_MS = Number(process.env.HARNESS_MD_SETTLE_MS ?? 120_000);
 const DEVICES = Math.max(2, Number(process.env.HARNESS_MD_DEVICES ?? 2));
 
 const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms));
-
-/** Messages of a given text prefix this bot's real code actually persisted. */
-const postsMatching = (bot: HarnessBot, prefix: string) =>
-  new Set(
-    bot.captured
-      .filter((m) => m.content?.type === 'post' && (m.content.text ?? '').startsWith(prefix))
-      .map((m) => m.content?.text as string)
-  );
-
-/**
- * WHICH numbered messages are missing, not just how many.
- *
- * The count alone cannot separate two very different bugs, and the first
- * multi-device run produced exactly the ambiguity: one device persisted 52 of 100
- * in BOTH directions while every frame arrived and nothing failed to decrypt.
- *
- *   a contiguous TAIL missing  ⇒ the device stopped processing at some moment
- *                                (a receive-pipeline stall)
- *   SCATTERED gaps             ⇒ per-message drops
- *   every OTHER one            ⇒ something in dedupe/ordering
- *
- * Those need different investigations, so the run has to say which it is.
- */
-function missingReport(bot: HarnessBot, prefix: string, rounds: number): string {
-  const got = new Set<number>();
-  for (const text of postsMatching(bot, prefix)) {
-    const n = Number(/#(\d+)$/.exec(text)?.[1]);
-    if (Number.isFinite(n)) got.add(n);
-  }
-  const missing: number[] = [];
-  for (let i = 1; i <= rounds; i++) if (!got.has(i)) missing.push(i);
-  if (missing.length === 0) return 'none';
-
-  // Contiguous tail? i.e. everything from some point on is absent.
-  const isTail = missing[missing.length - 1] === rounds && missing.length === rounds - missing[0] + 1;
-  const shape = isTail
-    ? `CONTIGUOUS TAIL from #${missing[0]} — looks like the device STOPPED`
-    : `scattered (${missing.length} gaps, first #${missing[0]}, last #${missing[missing.length - 1]})`;
-  const sample = missing.slice(0, 24).join(',') + (missing.length > 24 ? ',…' : '');
-  return `${shape}  missing=[${sample}]`;
-}
 
 test(
   'dm-multidevice: an N-device account, per-device arrival',
